Move ColorPicker to editor-toolbar
