+++ ts/editor/editor-toolbar/ColorPicker.svelte
